chore: decouple earn table
