+++ src/components/token-logo/StackTokenLogo.tsx
@@ -16,7 +16,11 @@ const StackTokenLogo = React.memo(({ tokens, sx = {}, ...props }: Props) => {
       {tokens.map((token, index) => {
         if (token.symbol === 'FRAXBP') {
           return (
-            <Box variant="layout.verticalAlign" key={token.address}>
+            <Box
+              variant="layout.verticalAlign"
+              sx={{ width: '28px' }}
+              key={token.address}
+            >
               <TokenLogo
                 sx={{
                   position: 'relative',
@@ -36,12 +40,19 @@ const StackTokenLogo = React.memo(({ tokens, sx = {}, ...props }: Props) => {
         }
 
         return (
-          <TokenLogo
-            key={token.address}
-            sx={{ position: 'relative', left: index ? `${-6 * index}px` : 0 }}
-            symbol={token.symbol}
-            src={token.logo}
-          />
+          <Box
+            sx={{
+              position: 'relative',
+              left: index ? `${-6 * index}px` : 0,
+              marginRight: index ? `${-6 * index}px` : 0,
+            }}
+          >
+            <TokenLogo
+              key={token.address}
+              symbol={token.symbol}
+              src={token.logo}
+            />
+          </Box>
         )
       })}
     </Box>
