+++ src/views/earn/hooks/useEarnTableColumns.tsx
@@ -2,23 +2,22 @@ import { t } from '@lingui/macro'
 import { createColumnHelper } from '@tanstack/react-table'
 import Help from 'components/help'
 import ChainLogo from 'components/icons/ChainLogo'
-import ExternalArrowIcon from 'components/icons/ExternalArrowIcon'
+import Aerodrome from 'components/icons/logos/Aerodrome'
+import Balancer from 'components/icons/logos/Balancer'
 import Convex from 'components/icons/logos/Convex'
 import Curve from 'components/icons/logos/Curve'
+import Extra from 'components/icons/logos/Extra'
+import Stakedao from 'components/icons/logos/Stakedao'
+import Uniswap from 'components/icons/logos/Uniswap'
 import Yearn from 'components/icons/logos/Yearn'
 import StackTokenLogo from 'components/token-logo/StackTokenLogo'
+import mixpanel from 'mixpanel-browser'
 import React, { useMemo } from 'react'
 import { Pool } from 'state/pools/atoms'
 import { Box, Text } from 'theme-ui'
 import { formatCurrency } from 'utils'
 import { ChainId } from 'utils/chains'
 import { CHAIN_TAGS, LP_PROJECTS } from 'utils/constants'
-import mixpanel from 'mixpanel-browser'
-import Stakedao from 'components/icons/logos/Stakedao'
-import Uniswap from 'components/icons/logos/Uniswap'
-import Balancer from 'components/icons/logos/Balancer'
-import Aerodrome from 'components/icons/logos/Aerodrome'
-import Extra from 'components/icons/logos/Extra'
 
 const chainMap: Record<string, number> = {
   Ethereum: ChainId.Mainnet,
@@ -70,10 +69,9 @@ const useEarnTableColumns = () => {
               }}
             >
               <StackTokenLogo tokens={data.row.original.underlyingTokens} />
-              <Text mx="1" sx={{ textDecoration: 'underline' }}>
+              <Text ml="2" sx={{ textDecoration: 'underline' }}>
                 {data.getValue()}
               </Text>
-              <ExternalArrowIcon />
             </Box>
           )
         },
